+++ data-prepper-logstash-configuration/src/main/java/org/opensearch/dataprepper/logstash/mapping/mutate/MutateMapper.java
@@ -3,9 +3,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.logstash.mapping;
+package org.opensearch.dataprepper.logstash.mapping.mutate;
 
 import com.amazon.dataprepper.model.configuration.PluginModel;
+import org.opensearch.dataprepper.logstash.mapping.LogstashAttributesMappings;
+import org.opensearch.dataprepper.logstash.mapping.LogstashPluginAttributesMapper;
 import org.opensearch.dataprepper.logstash.mapping.mutate.AbstractConversion;
 import org.opensearch.dataprepper.logstash.mapping.mutate.AddEntryConversion;
 import org.opensearch.dataprepper.logstash.mapping.mutate.CopyValueConversion;
@@ -23,7 +25,7 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 
-class MutateMapper implements LogstashPluginAttributesMapper {
+public class MutateMapper implements LogstashPluginAttributesMapper {
     private final Map<String, AbstractConversion> conversionMap = new HashMap<>();
 
     public MutateMapper() {
