+++ data-prepper-logstash-configuration/src/main/java/org/opensearch/dataprepper/logstash/mapping/mutate/SubMutateAction.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.logstash.mapping;
+package org.opensearch.dataprepper.logstash.mapping.mutate;
 
 import com.amazon.dataprepper.model.configuration.PluginModel;
 import org.opensearch.dataprepper.logstash.model.LogstashAttribute;
