+++ data-prepper-logstash-configuration/src/main/java/org/opensearch/dataprepper/logstash/mapping/mutate/UppercaseStringConversion.java
@@ -7,7 +7,7 @@ package org.opensearch.dataprepper.logstash.mapping.mutate;
 
 import java.util.List;
 
-public class UppercaseStringConversion extends AbstractConversion<String> {
+class UppercaseStringConversion extends AbstractConversion<String> {
     public static String getLogstashName() {
         return "uppercase";
     }
