+++ data-prepper-logstash-configuration/src/main/resources/org/opensearch/dataprepper/logstash/mapping/mutate.mapping.yaml
@@ -1,1 +1,1 @@
-customPluginMapperClass: org.opensearch.dataprepper.logstash.mapping.MutateMapper
+customPluginMapperClass: org.opensearch.dataprepper.logstash.mapping.mutate.MutateMapper
